requirements.txt
fastapi
uvicorn
pymongo
chromadb
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
pydantic
google-generativeai
